+++ lib/moco/entities/project.rb
@@ -2,16 +2,21 @@
 
 module MOCO
   class Project < BaseEntity
-    def customers
-      # NOTE: customer_id is not directly available in Project attributes,
-      # need to use the customer association.
-      customer_obj = association(:customer)
-      customer_obj ? client.companies.find(customer_obj.id) : nil
+    def customer
+      # Use the association method to fetch the customer
+      association(:customer, "Company")
     end
 
     # Fetches activities associated with this project.
     def activities
-      client.activities.where(project_id: id)
+      # Use the has_many method to fetch activities
+      has_many(:activities)
+    end
+
+    # Fetches tasks associated with this project.
+    def tasks
+      # Use the has_many method to fetch tasks
+      has_many(:tasks)
     end
 
     def active?